Remove StorageComponent, the functionality now resides in StepCounterComponent Fixed bug that caused steps to go up when going between activities, but lost the ability to keep track of steps if phone was restarted. Add calories, they get shown and calculated in StepCounterComponent.
